+++ src/engine/audio/alaudioasset.cpp
@@ -36,7 +36,10 @@ std::unique_ptr<AudioAsset> LoadOgg(std::ifstream& input) {
     int numSamples = stb_vorbis_decode_memory((unsigned char*)(data), size, &channels, &sample_rate, &buffer);
 
     ALenum format = channels == 1 ? AL_FORMAT_MONO16 : AL_FORMAT_STEREO16;
-    alBufferData(audio_asset->buffer, format, buffer, numSamples * channels * sizeof(int16), sample_rate);
+    alBufferData(
+        audio_asset->buffer, format, buffer,
+        static_cast<unsigned long long>(numSamples) * static_cast<unsigned long long>(channels) * sizeof(int16),
+        sample_rate);
     audio_asset->length = static_cast<float>(numSamples) / static_cast<float>(sample_rate);
 
     // Free memory
@@ -52,7 +55,10 @@ std::unique_ptr<AudioAsset> LoadOgg(uint8_t* buffer, int size) {
     int sample_rate;
     int num_samples = stb_vorbis_decode_memory(buffer, size, &channels, &sample_rate, &output);
     ALenum format = (channels == 1) ? AL_FORMAT_MONO16 : AL_FORMAT_STEREO16;
-    alBufferData(audio_asset->buffer, format, output, num_samples * channels * sizeof(int16), sample_rate);
+    alBufferData(
+        audio_asset->buffer, format, output,
+        static_cast<unsigned long long>(num_samples) * static_cast<unsigned long long>(channels) * sizeof(int16),
+        sample_rate);
     audio_asset->length = static_cast<float>(num_samples) / static_cast<float>(sample_rate);
 
     // Free memory
